+++ crates/stackable-versioned-macros/src/attrs/container.rs
@@ -53,8 +53,9 @@ pub struct ContainerSkipArguments {
 /// - `scale`: Configure the scale subresource for horizontal pod autoscaling integration.
 /// - `shortname`: Set a shortname for the CR object. This can be specified multiple
 ///   times.
-/// - `doc`: Override the root description of the generated CRD. If not set, kube
-///   generates a generic one automatically.
+/// - `doc`: Set the root description of the generated CRD. This argument is required, so
+///   that every CRD has a meaningful description instead of the generic one that kube
+///   generates by default.
 /// - `skip`: Controls skipping parts of the generation.
 #[derive(Clone, Debug, FromMeta)]
 pub struct StructCrdArguments {
@@ -73,7 +74,7 @@ pub struct StructCrdArguments {
     pub shortnames: Vec<String>,
     // category
     // selectable
-    pub doc: Option<String>,
+    pub doc: String,
     // annotation
     // label
 }